Move the potential solvers to solver.cpp
